- Update Mascot validation helper to use MS query properties - Add support for PeptideMatchProperties in all required classes - Store peptide match properties in the SQLite storer

+++ Proline-OM/src/main/scala/fr/proline/core/om/model/msi/PropertiesClasses.scala
@@ -22,7 +22,7 @@ case class PeptideMatchProperties (
 
 @JsonSnakeCase
 case class PeptideMatchMascotProperties (
-  @BeanProperty var expectationValue: Float,
+  @BeanProperty var expectationValue: Double,
   @BeanProperty var readableVarMods: Option[String] = None,
   @BeanProperty var varModsPositions: Option[String] = None
 )
